removed Helper class
